+++ nes/src/RingBuffer.cpp
@@ -1,7 +1,7 @@
 #include "RingBuffer.h"
 
 RingBuffer::RingBuffer(int size) : mSize(size), mUnreadCount(0) {
-  mStart = new float[mSize]{0};
+  mStart = new double[mSize]{0};
   mEnd = mStart + size;
   readPointer = mStart;
   writePointer = mEnd;
@@ -9,7 +9,7 @@ RingBuffer::RingBuffer(int size) : mSize(size), mUnreadCount(0) {
 
 RingBuffer::~RingBuffer() { delete[] mStart; }
 
-void RingBuffer::Write(float data) {
+void RingBuffer::Write(double data) {
   std::unique_lock<std::mutex> lock(mGuard);
   condition.wait(lock, [&]() { return mUnreadCount < mSize; });
   *writePointer = data;
@@ -17,7 +17,7 @@ void RingBuffer::Write(float data) {
   mUnreadCount++;
 }
 
-float RingBuffer::Read() {
+double RingBuffer::Read() {
   std::unique_lock<std::mutex> lock(mGuard);
   if (mUnreadCount > 0) {
     mLastRead = *readPointer;
@@ -28,7 +28,7 @@ float RingBuffer::Read() {
   return mLastRead;
 }
 
-float* RingBuffer::Advance(float* current) {
+double* RingBuffer::Advance(double* current) {
   if (current == mEnd) {
     return mStart;
   }
